Make maybeEnqueueFontStyle into a hook.
