+++ assets/src/edit-story/app/font/actions/useLoadFontFiles.js
@@ -1,0 +1,60 @@
+/**
+ * External dependencies
+ */
+import PropTypes from 'prop-types';
+
+
+import { useCallback } from '@wordpress/element';
+
+
+function useLoadFontFiles( { getFontByName } ) {
+	/**
+	 * Adds a <link> element to the <head> for a given font in case there is none yet.
+	 *
+	 * Allows dynamically enqueuing font styles when needed.
+	 *
+	 * @param {string} name Font name.
+	 */
+	const maybeEnqueueFontStyle = useCallback( ( name ) => {
+		if ( ! name ) {
+			return;
+		}
+
+		const font = getFontByName( name );
+		if ( ! font ) {
+			return;
+		}
+
+		const { handle, src } = font;
+		if ( ! handle || ! src ) {
+			return;
+		}
+
+		const element = document.getElementById( handle );
+
+		if ( element ) {
+			return;
+		}
+
+		const fontStylesheet = document.createElement( 'link' );
+		fontStylesheet.id = handle;
+		fontStylesheet.href = src;
+		fontStylesheet.rel = 'stylesheet';
+		fontStylesheet.type = 'text/css';
+		fontStylesheet.media = 'all';
+		fontStylesheet.crossOrigin = 'anonymous';
+
+		document.head.appendChild( fontStylesheet );
+	}, [ getFontByName ]);
+
+	return maybeEnqueueFontStyle;
+}
+
+useLoadFontFiles.propTypes = {
+	children: PropTypes.oneOfType( [
+		PropTypes.arrayOf( PropTypes.node ),
+		PropTypes.node,
+	] ).isRequired,
+};
+
+export default useLoadFontFiles;
