+++ assets/src/edit-story/elements/text/util.js
@@ -44,43 +44,3 @@ export const getHandleKeyCommand = ( setEditorState ) => ( command, currentEdito
 	}
 	return 'not-handled';
 };
-
-/**
- * Adds a <link> element to the <head> for a given font in case there is none yet.
- *
- * Allows dynamically enqueuing font styles when needed.
- *
- * @param {string} name Font name.
- * @param {Function} getFontByName Function to get current font.
- */
-export const maybeEnqueueFontStyle = ( name, getFontByName ) => {
-	if ( ! name ) {
-		return;
-	}
-
-	const font = getFontByName( name );
-	if ( ! font ) {
-		return;
-	}
-
-	const { handle, src } = font;
-	if ( ! handle || ! src ) {
-		return;
-	}
-
-	const element = document.getElementById( handle );
-
-	if ( element ) {
-		return;
-	}
-
-	const fontStylesheet = document.createElement( 'link' );
-	fontStylesheet.id = handle;
-	fontStylesheet.href = src;
-	fontStylesheet.rel = 'stylesheet';
-	fontStylesheet.type = 'text/css';
-	fontStylesheet.media = 'all';
-	fontStylesheet.crossOrigin = 'anonymous';
-
-	document.head.appendChild( fontStylesheet );
-};
